# Conflicts:
#	Calin/Assets/Script/Weapons/Shotgun/Shotgun Base.asset

+++ Calin/Assets/Script/Ennemis/Rush/SO_Rush.cs
@@ -8,10 +8,16 @@ public class SO_Rush : ScriptableObject
     [Header("Rush Spec")] 
     
     public int levelPlayer;
+    
     public List<float> life;
     public List<float> velocity;
     public List<float> numberBullet;
-    public List<float> xp;
+    public List<int> xp;
+    
+    
+    
+    
+    
     
     public List<float> timeCooldown;
     public int timeCooldownIndex;
